requirements.txt

dagster
dagster-webserver
dbt-mysql
pandas
sqlalchemy
pymysql
requests
kafka-python
pyspark==3.5.0
